requirements.txt
numpy==1.20.1
pandas
matplotlib
scikit-image
scipy
cffi
imageio
tqdm
tensorflow
tensorboardX
yacs
pyyaml
easydict
cython
mmcv==0.4.4
ipython
thop
# torch>=1.2
torchvision>=0.4.0
pypng
# spatial-correlation-sampler
